# sera-back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sera API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/projects/init": {
            "post": {
                "tags": [
                    "Step"
                ],
                "summary": "Init project",
                "description": "Init project",
                "operationId": "InitProject",
                "requestBody": {
                    "description": "Init project",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "project_request_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Project title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Project description"
                                    },
                                    "project_request_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Project title"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Project description"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "ongoing"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-05-05T14:48:00.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-05-05T14:48:00.000000Z"
                                        },
                                        "project_request_id": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Project request already processed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project request not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Project request not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/projects/{id}/steps": {
            "get": {
                "tags": [
                    "Step"
                ],
                "summary": "Get project steps",
                "description": "Get project steps",
                "operationId": "GetSteps",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status of the steps",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ongoing",
                                "done",
                                "not_started"
                            ]
                        }
                    },
                    {
                        "name": "step",
                        "in": "query",
                        "description": "Select a specific step",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Planning",
                                "Capture",
                                "Post-Production",
                                "Transcription",
                                "Subtitling",
                                "Editorial"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project steps"
                    },
                    "404": {
                        "description": "Project not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Project not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Step not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}